LayoutTests/fast/page-color-sampling/color-sampling-header-visibility-expected.txt

PASS colorBeforeHiding is "rgb(255, 0, 0)"
PASS colorAfterHiding is "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
